feat(mm-next/topic-list): move GPT ad after the 12th article
